Refactor DirectX and Windows compatibility layers for macOS port
- Replaced ddraw.h inclusion with a custom compatibility header for non-Windows systems.
- Added DirectX 8 compatibility headers (d3d8caps.h, d3d8types.h) to support macOS.
- Implemented Windows API compatibility functions in windows.h and win32_compat.h.
- Refactored shader.h to handle macro conflicts with system headers.
- Updated README.md to reflect progress on macOS support and OpenGL integration.
- Enhanced testing documentation for macOS and Linux compatibility.
- Improved forward declarations in texture and textureloader headers to resolve compilation issues.
- Removed unnecessary mmsystem.h inclusion from dx8caps.cpp and related files.
- Added compatibility for DirectDraw structures and constants.
- Documented macOS port progress in MACOS_PORT.md, highlighting completed components and ongoing efforts.

+++ Generals/Code/Libraries/Source/WWVegas/WW3D2/assetmgr.cpp
@@ -74,6 +74,8 @@
 
 #include "assetmgr.h"
 #include <assert.h>
+#include "win32_compat.h"
+#include <cstddef>
 
 #include "bittype.h"
 #include "chunkio.h"
@@ -798,7 +800,7 @@ RenderObjClass * WW3DAssetManager::Create_Render_Obj(const char * name)
 		char filename [MAX_PATH];
 		const char *mesh_name = ::strchr (name, '.');
 		if (mesh_name != NULL) {
-			::lstrcpyn (filename, name, ((int)mesh_name) - ((int)name) + 1);
+			::lstrcpyn (filename, name, ((ptrdiff_t)mesh_name) - ((ptrdiff_t)name) + 1);
 			::lstrcat (filename, ".w3d");
 		} else {
 			sprintf( filename, "%s.w3d", name);
